Move the Endless event processors to a common module
This will be useful as it provides a place to put various helpers we
will need to share between them.
