+++ azafea/event_processors/endless/activation/tests/integration/test_activation_v1.py
@@ -14,10 +14,10 @@ from azafea.tests.integration import IntegrationTest
 
 
 class TestActivation(IntegrationTest):
-    handler_module = 'azafea.event_processors.activation.v1'
+    handler_module = 'azafea.event_processors.endless.activation.v1'
 
     def test_activation_v1(self):
-        from azafea.event_processors.activation.v1.handler import Activation
+        from azafea.event_processors.endless.activation.v1.handler import Activation
 
         # Create the table
         self.run_subcommand('initdb')
