+++ azafea/event_processors/endless/metrics/tests/test_events.py
@@ -18,7 +18,8 @@ from sqlalchemy.types import Integer, Unicode
 
 
 def test_singular_registry():
-    from azafea.event_processors.metrics.events._base import SINGULAR_EVENT_MODELS, SingularEvent
+    from azafea.event_processors.endless.metrics.events._base import (SINGULAR_EVENT_MODELS,
+                                                                      SingularEvent)
 
     class TestSingularEvent(SingularEvent):
         __tablename__ = 'test_singular'
@@ -28,7 +29,8 @@ def test_singular_registry():
 
 
 def test_aggregate_registry():
-    from azafea.event_processors.metrics.events._base import AGGREGATE_EVENT_MODELS, AggregateEvent
+    from azafea.event_processors.endless.metrics.events._base import (AGGREGATE_EVENT_MODELS,
+                                                                      AggregateEvent)
 
     class TestAggregateEvent(AggregateEvent):
         __tablename__ = 'test_aggregate'
@@ -38,7 +40,8 @@ def test_aggregate_registry():
 
 
 def test_sequence_registry():
-    from azafea.event_processors.metrics.events._base import SEQUENCE_EVENT_MODELS, SequenceEvent
+    from azafea.event_processors.endless.metrics.events._base import (SEQUENCE_EVENT_MODELS,
+                                                                      SequenceEvent)
 
     class TestSequenceEvent(SequenceEvent):
         __tablename__ = 'test_sequence'
@@ -48,7 +51,7 @@ def test_sequence_registry():
 
 
 def test_new_event_no_payload():
-    from azafea.event_processors.metrics.events._base import SingularEvent
+    from azafea.event_processors.endless.metrics.events._base import SingularEvent
 
     class TestEventNoPayload(SingularEvent):
         __tablename__ = 'test_singular_no_payload'
@@ -60,7 +63,7 @@ def test_new_event_no_payload():
 
 
 def test_new_event_with_payload():
-    from azafea.event_processors.metrics.events._base import SingularEvent
+    from azafea.event_processors.endless.metrics.events._base import SingularEvent
 
     class TestEventWithPayload(SingularEvent):
         __tablename__ = 'test_singular_with_payload'
@@ -85,7 +88,7 @@ def test_new_event_with_payload():
 
 def test_new_event_no_payload_but_payload_given(capfd):
     from azafea.logging import setup_logging
-    from azafea.event_processors.metrics.events._base import SingularEvent
+    from azafea.event_processors.endless.metrics.events._base import SingularEvent
 
     setup_logging(verbose=False)
 
@@ -103,7 +106,8 @@ def test_new_event_no_payload_but_payload_given(capfd):
 
 
 def test_new_event_no_payload_given():
-    from azafea.event_processors.metrics.events._base import EmptyPayloadError, SingularEvent
+    from azafea.event_processors.endless.metrics.events._base import (EmptyPayloadError,
+                                                                      SingularEvent)
 
     class TestEventNoPayloadGiven(SingularEvent):
         __tablename__ = 'test_singular_no_payload_given'
@@ -120,7 +124,8 @@ def test_new_event_no_payload_given():
 
 
 def test_new_event_wrong_payload_given():
-    from azafea.event_processors.metrics.events._base import SingularEvent, WrongPayloadError
+    from azafea.event_processors.endless.metrics.events._base import (SingularEvent,
+                                                                      WrongPayloadError)
 
     class TestEventWrongPayloadGiven(SingularEvent):
         __tablename__ = 'test_singular_wrong_payload_given'
@@ -137,7 +142,7 @@ def test_new_event_wrong_payload_given():
 
 
 def test_new_unknown_event():
-    from azafea.event_processors.metrics.events._base import UnknownEvent
+    from azafea.event_processors.endless.metrics.events._base import UnknownEvent
 
     class TestUnknownEvent(UnknownEvent):
         __tablename__ = 'test_unknown'
@@ -373,7 +378,7 @@ def test_new_unknown_event():
     ('WindowsLicenseTables', GLib.Variant('u', 0), {'tables': 0}),
 ])
 def test_singular_event(event_model_name, payload, expected_attrs):
-    from azafea.event_processors.metrics import events
+    from azafea.event_processors.endless.metrics import events
 
     event_model = getattr(events, event_model_name)
     maybe_payload = GLib.Variant('mv', payload)
@@ -385,7 +390,7 @@ def test_singular_event(event_model_name, payload, expected_attrs):
 
 
 def test_invalid_hack_clubhouse_progress_event():
-    from azafea.event_processors.metrics.events import HackClubhouseProgress
+    from azafea.event_processors.endless.metrics.events import HackClubhouseProgress
 
     # Make an invalid payload with missing keys
     payload = GLib.Variant('mv', GLib.Variant('a{sv}', {'complete': GLib.Variant('b', True)}))
@@ -402,7 +407,7 @@ def test_hack_clubhouse_progress_event_with_unknown_key():
     # Additional event attributes passed to the constructor are ignored by our base model, but for
     # this specific event the code parsing the payload does its own ignoring as well, so let's test
     # it to be sure.
-    from azafea.event_processors.metrics.events import HackClubhouseProgress
+    from azafea.event_processors.endless.metrics.events import HackClubhouseProgress
 
     # Make an invalid payload with extra keys
     payload = GLib.Variant('mv', GLib.Variant('a{sv}', {
@@ -430,7 +435,7 @@ def test_hack_clubhouse_progress_event_with_unknown_key():
     ('UserIsLoggedIn', GLib.Variant('u', 1000), {'logged_in_user_id': 1000}),
 ])
 def test_sequence_event(event_model_name, payload, expected_attrs):
-    from azafea.event_processors.metrics import events
+    from azafea.event_processors.endless.metrics import events
 
     event_model = getattr(events, event_model_name)
     maybe_payload = GLib.Variant('mv', payload)
