+++ azafea/event_processors/endless/ping/tests/test_ping_validation.py
@@ -11,25 +11,25 @@ import pytest
 
 
 def test_valid_country():
-    from azafea.event_processors.activation.v1.handler import Activation
+    from azafea.event_processors.endless.ping.v1.handler import Ping
 
-    activation = Activation(country='HKG')
-    assert activation.country == 'HKG'
+    ping = Ping(country='HKG')
+    assert ping.country == 'HKG'
 
 
 @pytest.mark.parametrize('country', ['', None])
 def test_empty_country(country):
-    from azafea.event_processors.activation.v1.handler import Activation
+    from azafea.event_processors.endless.ping.v1.handler import Ping
 
-    activation = Activation(country=country)
-    assert activation.country is None
+    ping = Ping(country=country)
+    assert ping.country is None
 
 
 @pytest.mark.parametrize('country', ['HK', 'Hong Kong'])
 def test_invalid_country(country):
-    from azafea.event_processors.activation.v1.handler import Activation
+    from azafea.event_processors.endless.ping.v1.handler import Ping
 
     with pytest.raises(ValueError) as excinfo:
-        Activation(country=country)
+        Ping(country=country)
 
     assert f'country has wrong length: {country}' in str(excinfo.value)
